refactor: Refactor several files based on change 'ID & UserID' type in domain layer

+++ internal/domain/dto/servicedto/usertaskservicedto/create_task.go
@@ -5,7 +5,7 @@ import (
 )
 
 type CreateTaskRequest struct {
-	UserID      uint
+	UserID      uint64
 	Title       string
 	Description string
 	Status      entity.TaskStatus
